requirements-test.txt
# Testing dependencies for Medical Text Classification project

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0  # For parallel test execution
pytest-html>=3.1.0   # For HTML test reports
pytest-json-report>=1.5.0  # For JSON test reports

# HTTP testing
httpx>=0.24.0
requests>=2.28.0

# Database testing
pytest-postgresql>=4.1.0  # For PostgreSQL testing
sqlalchemy-utils>=0.40.0  # Database utilities for testing

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.60.0
psutil>=5.9.0

# Code quality and linting
flake8>=5.0.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0
bandit>=1.7.0  # Security linting

# Coverage reporting
coverage[toml]>=7.0.0

# Mocking and fixtures
factory-boy>=3.2.0  # For creating test data
faker>=18.0.0       # For generating fake data
responses>=0.23.0   # For mocking HTTP requests

# API testing
fastapi[all]>=0.95.0  # Includes TestClient
starlette>=0.27.0

# Additional utilities
freezegun>=1.2.0    # For mocking datetime
pytest-env>=0.8.0   # For environment variable management
pytest-timeout>=2.1.0  # For test timeouts
